combine owned tokens and all tokens
